fix early move attempt. better message logs

+++ src/DownloadWatcher.Application/Application.cs
@@ -1,3 +1,4 @@
+
 namespace DownloadWatcher.Application;
 
 public class Application
@@ -17,17 +18,39 @@ public class Application
         }
         try
         {
+            FileInfo fileInfo = new(path);
+            if (IsFileStillDownloading(fileInfo))
+            {
+                return;
+            }
+
             DirectoryInfo? parent = Directory.GetParent(newName);
             if (parent != null && !Path.Exists(parent.FullName))
             {
                 Directory.CreateDirectory(parent.FullName);
             }
-            Console.WriteLine($"Moving file to {newName}");
+            Console.WriteLine($"[{NowString()}] Moving file to {newName}");
             File.Move(path, newName);
+        }
+        catch (IOException)
+        {
+            File.Delete(path);
         }
         catch (Exception e)
         {
-            Console.WriteLine($"Error moving file: {e.Message}");
+            Console.WriteLine($"[{NowString()}] Error moving file: {e.Message}");
         }
+    }
+
+    static string NowString()
+    {
+        return DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss");
+    }
+    static bool IsFileStillDownloading(FileInfo fileInfo)
+    {
+        long initialSize = fileInfo.Length;
+        Thread.Sleep(1500);
+        fileInfo.Refresh();
+        return fileInfo.Length != initialSize;
     }
 }
